:recycle: :boom: Unify hooks into a single MlflowHook class (#315)
