+++ kedro_mlflow/framework/hooks/__init__.py
@@ -1,2 +1,1 @@
-from .node_hook import MlflowNodeHook, mlflow_node_hook
-from .pipeline_hook import MlflowPipelineHook, mlflow_pipeline_hook
+from .mlflow_hook import MlflowHook, mlflow_hook
